Commit before making Unpack return immutable components.
There will also be an UnpackMutable method for components you wish to mutate.
The way we achieve this is by wrapping the component returned from Unpack in a struct called Immutable<T> that will contain the component as readonly.

Adding this to support lightweight component previous state conservation, and to add support for an OnComponentSet handler.